# portfolio/static/assets/json/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Portfolio",
    "description": "APIs for my portfolio app",
    "contact": {
      "name": "Sean Jones",
      "url": "seanjones.app",
      "email": "sean.d.jones92@gmail.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "/api",
      "variables": {}
    }
  ],
  "paths": {
    "/healthcheck": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "healthcheck",
        "description": "Check that the application is up and running",
        "operationId": "healthcheck",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
    "/resume": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "resume",
        "description": "Return my resume in JSON format",
        "operationId": "resume",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
    "/mockme": {
      "post": {
        "tags": [
          "Projects"
        ],
        "summary": "mockme",
        "description": "Repeat the provided message in a mocking tone",
        "operationId": "mockme",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/mockmerequest"
              },
              "example": {
                "message": "why are you doing this?"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
    "/roll": {
      "post": {
        "tags": [
          "Projects"
        ],
        "summary": "roll",
        "description": "Roll and number of dice with any number of sides and apply a single modifier",
        "operationId": "roll",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/rollrequest"
              },
              "example": {
                "dicenumber": 1,
                "dicesides": 20,
                "rollmodifier": 0
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
    "/emacs": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "emacs",
        "description": "Return a bash script to setup my emacs config",
        "operationId": "emacs",
        "parameters": [
          {
            "name": "private",
            "in": "query",
            "description": "If private is set to true return a config that requires SSH trust to Github",
            "style": "form",
            "explode": true,
            "schema": {
              "type": "boolean",
              "example": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    }
  },
  "components": {
    "schemas": {
      "mockmerequest": {
        "title": "mockmerequest",
        "required": [
          "message"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "example": {
          "message": "why are you doing this?"
        }
      },
      "rollrequest": {
        "title": "rollrequest",
        "required": [
          "dicenumber",
          "dicesides",
          "rollmodifier"
        ],
        "type": "object",
        "properties": {
          "dicenumber": {
            "type": "integer",
            "format": "int32"
          },
          "dicesides": {
            "type": "integer",
            "format": "int32"
          },
          "rollmodifier": {
            "type": "integer",
            "format": "int32"
          }
        },
        "example": {
          "dicenumber": 1,
          "dicesides": 20,
          "rollmodifier": 0
        }
      }
    }
  },
  "tags": [
    {
      "name": "Misc",
      "description": ""
    }
  ]
}